+++ SBCK/tools/__stats.py
@@ -30,43 +30,6 @@ import scipy.special as scs
 ## Functions ##
 ###############
 
-def bin_width_estimator( X , method = "auto" ):##{{{
-	"""
-	SBCK.tools.bin_width_estimator
-	==============================
-	
-	Estimate the width of the bin to build an histogram of X
-	
-	Parameters
-	----------
-	X    : np.array[ shape = (n_samples,n_features) ] or list(np.array)
-		A dataset or a list of dataset X containing n_samples observations of n_features random variables.
-	method : string = [ "auto" , "Sturges" , "FD" ]
-		Method to estimate bin_width. If method == "auto", "Sturges" is selected if n_samples < 1000, else "FD"
-	
-	Returns
-	-------
-	bin_width : np.array[ shape = (n_features) ]
-		bin_width of each features.
-	"""
-	
-	if type(X) == list:
-		return np.min( [ bin_width_estimator( x , method ) for x in X ] , axis = 0 )
-	
-	if X.ndim == 1 : X = X.reshape(-1,1)
-	
-	if method == "auto":
-		method = "Sturges" if X.shape[0] < 1000 else "FD"
-	
-	if method == "Sturges":
-		nh = np.log2( X.shape[0] ) + 1.
-		bin_width = np.zeros(X.shape[1]) + 1. / nh
-	else:
-		bin_width = 2. * ( np.percentile( X , q = 75 , axis = 0 ) - np.percentile( X , q = 25 , axis = 0 ) ) / np.power( X.shape[0] , 1. / 3. )
-	
-	return bin_width
-##}}}
-
 def rvs_spd_matrix( dim ):##{{{
 	"""
 	TODO move to tools